+++ CHANGELOG.md
@@ -22,6 +22,8 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 - SC-9077 - refactored store actions returning data
 - SC-9126 - changed the way the stores state is accessed; `mapState` and `store.state` were substituted for `mapGetters` and `store.getters` respectively; and changed the naming of the template and several other stores getters
 - SC-9122 - changed getters logic from components to the store
+- SC-9108 - refactored news
+- SC-9168 - POC for Typescript store module (news)
 
 ### Fixed